third_party/WebKit/LayoutTests/http/tests/devtools/load-file-resource-for-frontend-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Test loading file resource from front-end 


Running: testResourceFromFileScheme
Loading resource from resources/file-resource-to-load-for-frontend.txt
Content: SUCCESS
